Rename App.tsx to app.tsx for consistency
