+++ apps/blank-app/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { App } from "./App";
+import { App } from "./app";
 import { NimbusProvider } from "@commercetools/nimbus";
 
 createRoot(document.getElementById("root")!).render(
